backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.0
gunicorn==21.2.0
